requirements.txt
can-isotp==2.0.7
numpy==2.3.3
packaging==25.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
python-can==4.6.1
typing_extensions==4.15.0
udsoncan==1.25.1
wrapt==1.17.3
